+++ service/src/test/java/eu/europa/ec/fisheries/uvms/movement/service/MockData.java
@@ -64,8 +64,8 @@ public class MockData {
         Point point = factory.createPoint(coordinate);
         point.setSRID(4326);
         movement.setLocation(point);
-        movement.setHeading(latlong.bearing);
-        movement.setSpeed(latlong.speed);
+        movement.setHeading((float)latlong.bearing);
+        movement.setSpeed((float)latlong.speed);
         movement.setMovementSource(MovementSourceType.NAF);
         movement.setStatus("TEST");
         movement.setUpdated(Instant.now());
